# growup-auth/src/main/java/com/growup/auth/infrastructure/adapter/web/dto/LoginRequest.java
package com.growup.auth.infrastructure.adapter.web.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * DTO para la petición de login.
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequest {
    private String email;
    private String password;
}