# specs/openapi.yaml
openapi: 3.0.0
info:
  title: API de GrowUp
  description: Especificación de la API para la plataforma de e-learning GrowUp.
  version: 1.1.0
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desarrollo Local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, TEACHER, STUDENT]
      description: Roles de usuario en el sistema
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - isActive
        - role
        - version
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del usuario
        name:
          type: string
          description: Nombre completo del usuario
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        isActive:
          type: boolean
          description: Indica si el usuario está activo (Baja lógica si es false)
        role:
          $ref: "#/components/schemas/Role"
        avatar:
          type: string
          format: uri
          description: URL del avatar del usuario
        joinDate:
          type: string
          format: date-time
          description: Fecha de registro del usuario
        bio:
          type: string
          description: Biografía del usuario (opcional)
        version:
          type: integer
          format: int64
          description: Versión de control de concurrencia
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Correo electrónico del usuario
        password:
          type: string
          format: password
          description: Contraseña del usuario
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          description: Nombre completo del usuario
        email:
          type: string
          format: email
          description: Correo electrónico
        password:
          type: string
          format: password
          description: Contraseña
        role:
          $ref: "#/components/schemas/Role"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: Nombre completo del usuario
        email:
          type: string
          format: email
          description: Correo electrónico
        avatar:
          type: string
          format: uri
          description: URL del avatar
        bio:
          type: string
          description: Biografía (opcional, para profesores)
        password:
          type: string
          format: password
          description: Nueva contraseña (opcional)
        version:
          type: integer
          format: int64
          description: Versión de control de concurrencia
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT de autenticación
        user:
          $ref: "#/components/schemas/User"
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: Código de estado HTTP
        message:
          type: string
          description: Mensaje de error
        timestamp:
          type: string
          format: date-time
          description: Momento del error

  # security: ELIMINADO - Se define seguridad por endpoint (ver paths)
paths:
...
  /auth/login:
    post:
      summary: Iniciar sesión
      description: Autenticar usuario con email y contraseña
      tags: [Autenticación]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      summary: Registrar nuevo usuario
      description: Crear una nueva cuenta de usuario
      tags: [Autenticación]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Datos inválidos o email ya registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      summary: Obtener usuario actual
      description: Obtener información del usuario autenticado
      tags: [Autenticación]
      responses:
        "200":
          description: Información del usuario actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Actualizar perfil propio
      description: Actualizar nombre, bio, avatar o contraseña del usuario autenticado
      tags: [Autenticación]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # Gestión de Usuarios (Admin)
  # ========================================
  /admin/users:
    get:
      summary: Listar todos los usuarios
      description: Obtener lista paginada de usuarios (Solo ADMIN)
      tags: [Admin]
      parameters:
        - in: query
          name: role
          required: false
          schema:
            $ref: "#/components/schemas/Role"
          description: Filtrar por rol
        - in: query
          name: isActive
          required: false
          schema:
            type: boolean
          description: Filtrar por estado (activo/inactivo)
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: No autorizado (requiere rol ADMIN)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    get:
      summary: Obtener detalles de usuario (Admin)
      description: Obtener información completa de cualquier usuario por ID
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      responses:
        "200":
          description: Detalles del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Modificar usuario (Admin)
      description: Modificar datos de cualquier usuario (incluyendo rol)
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Usuario modificado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/status:
    patch:
      summary: Cambiar estado de usuario (Alta/Baja Lógica)
      description: Activar o desactivar un usuario sin eliminarlo de la base de datos
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
                  description: Nuevo estado (true = activo, false = inactivo/baja lógica)
      responses:
        "200":
          description: Estado del usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # Gestión de Cursos
  # ========================================
...